add listing & details for product category

+++ Halwani.Core/ViewModels/CategoryModels/SubCateogryModel.cs
@@ -2,7 +2,7 @@
 {
     public class SubCateogryModel
     {
-        public int? SubCategoryId { get; set; }
+        public long? SubCategoryId { get; set; }
         public double? Goal { get; set; }
         public bool IsDeleted { get; set; }
         public string SubCategoryName { get; set; }
